dma-buf: include exp_name to anon file name
exp_name is introduced for debugging purpose. And we often list the
open file descriptors under /proc/<pid>/fd to find the list of anon
files of dma-buf optained by the process of <pid>. However it is not
possible to associate a specific open file descriptor listed in
/proc/<pid>/fd with the buffer list scrapped from
debugfs/dma-buf/bufinfo.

Attaching exp_name to the anon file name of dma-buf helps identifying
the exported module of the the buffers listed in /proc/<pid>/fd.

dma_buf.exp_name also has the copy of the string instead of pointing
where the string given by the exporter. This gives the exporter chances
to set different exp_name to the different buffers. Therefore we can
identify exact buffers from /proc/<pid>/fd and find the properties of
the buffers. The properties are defined by the exporter and never
exported elsewhere.

Change-Id: If35cf503ac24dea92eb50778e6aa50830145e24c
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>