[COMMON] serial: samsung: Move INTP clear phrase
Previously interrupt pending bit is cleared with 2 alternative
condition.
a. after transfer fifo size amount of data
b. after fifo count value is 0

but in this case, when the amount of data is bigger than fifo size,
it is possible to clear both the first INTP and the second INTP.

So now INTP will be cleard only when fifo count is 0 not to
miss any incoming INTP.

Change-Id: I69317c78501e36f8cc26fac21993dfd719538d01
Signed-off-by: Kyungwoo Kang <kwoo.kang@samsung.com>