iommu/exynos: fix handling masters without domain
Master devices that need an IO virtual address space should be
reigstered to 'domain-clients' attribute of a node with compatible
"samsung,exynos-iommu-bus". If phandle to a master device is not
registered to the list of domain clients, attaching IOMMU domain does
not effect because IO address space is not prepared.

Howerver Exynos iommu driver does not carefully handle that exception:
Daniel reported the following problem:
[   11.055792]  [7:      swapper/0:    1] Unable to handle kernel NULL pointer dereference at virtual address 0000000000000060
[   11.055927]  [7:      swapper/0:    1] Mem abort info:
[   11.059428]  [7:      swapper/0:    1]   ESR = 0x96000005
[   11.064808]  [7:      swapper/0:    1]   Exception class = DABT (current EL), IL = 32 bits
[   11.073058]  [7:      swapper/0:    1]   SET = 0, FnV = 0
[   11.078434]  [7:      swapper/0:    1]   EA = 0, S1PTW = 0
[   11.083903]  [7:      swapper/0:    1] Data abort info:
... snipped ...
[   11.312862]  [7:      swapper/0:    1] pstate: 60400005 (nZCv daif +PAN -UAO)
[   11.319983]  [7:      swapper/0:    1] PC is at exynos_iommu_of_xlate+0xec/0x1cc
[   11.327357]  [7:      swapper/0:    1] LR is at exynos_iommu_of_xlate+0x50/0x1cc

It is caused by accessing vmm_data in the list of iommu domain clients
for a given master device which is not registered as a client.

Driver should report error when a given master device is not found from
the list of iommu domain clients instead of trying to access vmm_data
that is never configured.

Change-Id: I9c8075c4733283eb52c934f8c5bbd979b290d939
Reported-by: Daniel Mentz <danielmentz@google.com>
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>