iommu/exynos: never disable MMU
If System MMU version does not support blocking traffics, the driver
disables System MMU instead of configuring the mmu to 'blocked' state.
The reason why configuring 'blocked' state is to prevent unexpected
traffics to the slave from master devices.
However the latest System MMU versions discarded 'blocked' state
because it is robust enough against changing configurations to System
MMU during it is enabled. On the other hand, we lost a workaround to
unexpected traffics from master devices while System MMUs are disabled
for local power down.
Let's keep System MMU enabled during local power down instead of really
disabling it. If unexpected traffics are generated by master devices,
it resuts in one of the followings:
 - successfully access to the slave if the target address is still
   mapped in the page table. This is okay because it is allocated to
   the master device.
 - page fault is occurred.

It is still okay if the System MMU is not reset when local power down
is canceled. System MMU is then initialized by __sysmmu_enable_nocount()
while it is enabled. It is safe to initialize System MMUs deprecating
'blocked' when they are enabled.

TLB invalidation is required in __sysmmu_disable_nocount() if System
MMU is not disabled. It prevents translations with stale translation
data in TLB if the page table is changed after
__sysmmu_disable_nocount() is called. System MMU driver does not
invalidate TLB if System MMU is deactivated by calling
__sysmmu_disable_nocount() even though page table is modified.

Likewise, read & update to MMU_CFG in __sysmmu_init_config() is also
changed to initializing MMU_CFG. It is not guaranteed that MMU_CFG has
always its reset value in __sysmmu_init_config() when local power down
is canceled. Moreover, __sysmmu_init_config() does not consider the
current setting of MMU_CFG. So, read & update is not necessary here.

Change-Id: I9ddc94309067ebafe858b5f29f3fab9fca9153ff
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>