iommu/exynos: fix register offset calcuation
MMU_SEC_VM_REG() adds register offset based on the vmid to the register
base address. ',' is used instead of '+'. Clang does not report it
because the sentence is evaluated with the last expression follows the
last ','. GCC reported it as a warning/error because it may not
intended by the developers.

Change-Id: I089b91a86bf38ee937c47c157c1aba6a015fda48
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>