wt010 touch driver for use in r11
This replaces the wt030 driver (version 0x01030100) with the wt010
driver (version 0x0013). The wt030 driver was expected to work with
both parts but this appears not to be the case.

Also, this driver has support for updating the firmware via sysfs
so we can flash compatible firmware (versions 2.2.0.*) to those parts
which were erroneously flashed with an incompatible version (2.5.*.*).

The following changes were made from the Raydium reference driver:

 - Removed raydium_power_on and raydium_power_init
   (touch driver will not manage its own power state)
 - Changed device tree parsing code to match existing device tree
 - Removed pinctrl states, "suspend" and "release"
 - Added code to initialize the reset gpio at the same time as the
   interrupt gpio. Without this the device was just staying in
   reset forever
 - Get the interrupt number from the interrupt gpio rather than
   the 'interrupts' field of the device tree
 - Use request_firmware to read the firmware binary rather than
   calling filp_open
 - Removed codepath for loading firmware from a header file
   since we weren't provided with the corresponding header files
 - Cleaned up formatting, which didn't match Linux kernel style

Bug: 159821513
Bug: 153345485
Change-Id: I2472e9a84dc10268d347088e1d5b8dcf712a0f6b