dma-buf: add dependency with debugfs configuration
The dma-buf-trace purpose is to show the relationship
between process and buffer by debugfs for debug.
That could be used on memtrack for EGL buffer size of process.
If debugfs is not enabled, the trace of buffer is unncessary
because all information is shown by debugfs, so dma buf trace
has dependency with debugfs configuration.

Change-Id: I28d95103c96d405775fa6b3752119fd0884315c2
Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>