Smart Amp: Set SysClk
The PLL reference clocks and input frequencies can be set by calling
snd_soc_codec_set_sysclk() function with the appropriate parameters.

This is required to get the smart amp to play audio.

For now, the I2S sclk signal will be used as the reference clock.
This may be changed in the future when we switch to TDM.

Bug: 159811216
Change-Id: Ib6f17ad24ed60b58740667ccd38a857780ef48cb