Move Smart Amp from I2S0 to I2S1 in driver
To get audio to play on DEV2, the smart amp I2S port must be moved to
I2S1.

Using the multi-codec functionality, the smart amp and mcu mic have been
placed on the same I2S bus. Configuring the smart amp for TDM will break
the mcu mic which is currently configured for L/R I2S. b/162451686
describes this and what is needed to make that change in the future.

There is no intention to keep reverse compatibility with DEV1.

Bug: 162976573
Change-Id: I2c7ed77fec8c5c0ccec85232562fb6ab3895269c