Port S2MU00X battery driver to extra/drivers
The purpose of the file is to communicate with the various power_supply
drivers in the system and collect the information from them to pass up
to userspace as a "battery" power_supply object. Also, it takes action
during charging to set up the charging state to match the parameters of
the battery specified in the device tree.

We will need to make significant modifications to this example driver in
the future. To make it easier, port a copy of it over to extra/drivers.

Changes:
- Remove references to S2mu00x which is the name of the example driver
- Remove usb and ac power supply creation as we only need battery
- remove IFCONN notifier option

Bug: 162782657
Change-Id: Iabf9e047e616520233075d951d521bdfda23e431