Cleaning up unused BT left over flags
Those flags came with the initial kernel drop where the bluetooth module was built in-tree.
In our case, we build the bluetooth module out of tree and all the needed flags are set accordingly in the designed out of tree directory.

Removing those flags to clean up and avoid any confusion.

Bug: 164534735

Signed-off-by: mullerf <mullerf@google.com>
Change-Id: I1caa8336204c833baf743b1e1873c5e876f26a49