android: ion: deny protected buffer from non-secure heap
ION_FLAG_PROTECTED is ignored by non-secure heaps including the system
heap. It is okay because it does not affect the safety of the system but
the users who can't access the allocated buffers.
However, we have found the case that allocating protected buffers from
a non-secure heap causes the kernel oops when trying to allocate device
address of the heap.

SI-18669 (MCD)

Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>
Change-Id: Ib167e0ad5d2f765b11d19a2679dfb65400ba82d1