dma-buf: do trace independently with debugfs
dma-buf-trace do trace only if debugfs is enabled because
the purpose of trace is to inform memtrack of buffers
per process by debugfs. However trace is so useful when
out-of-memory happens. trace shows buffers per process
on notifier of out-of-memory, so we do trace independently
with debugfs and manage trace information regardless of debugfs.

Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>
Change-Id: I5de59d5a0749a0c2a209bdd8156bcd44cc463aa8