Add debug_log and metrics nanohub channels
The debug_log channel will be used to send debug data from the MCU.  We
do not intend for this channel to carry a high amount of traffic in
normal operation.  Those who need to send a large amount of data will be
asked to gate that with a flag.  In the future, this may be removed in
favor of the UART channel, but that depends on when we can make the UART
channel reliable and efficient.

The metrics channel will be used to send Westworld data from the MCU.
This channel will stay as a nanohub channel, regardless of UART
readiness, as we need to be able to wake up the AP and flush the metrics
data when the buffer on the MCU is full.

Bug: 169786407
Bug: 172672211
Bug: 172517033
Change-Id: I444a379e8dda59c75422f5df45aefe7c5840cac2