Add support for EVT audio routing
This CL add support for the new audio routing on EVT while maintaining
compatibility with pre-EVT hardware.

In order to achieve this, a devicetree property is used to decide
between the EVT and Pre-EVT sounds cards.

Test on both DEV2 and modified DEV2 with EVT routing
(audio_playback.sh was modified to choose between UAIF0 and UAIF1)
Test: adb shell /vendor/bin/factory/audio_playback.sh

Bug: 177103103
Change-Id: I3a77f495ef5cbc1b096e99f1cdc65044e75639d0