Add handling for TimeSync GPIO
Check if timesync feature supported by detecting if timesync GPIO is
defined in device tree with name "timesync-gpio". If supported, create
an irq handler for timesync featture to handle interrupt of timesync
GPIO and add a queue to store system time when there is an interrupt
comes from timesync GPIO. Also, create a file node for user
space to get the timestamp in queue.

Bug: 175836015
Test: make
Change-Id: I9ff3a6aa1d525d0872958458c4dffbd44fdb23c7