(REVERT ME): Add DHD_PNO_VAL log output for debugging purposes
Automation testing is seeing an issue where WiFi does not always
reconnect automatically wen BT connection is lost/disconnected. The
root cause has been traced back to something going wrong when the
frameworks sends the request for PNO start scan, but the drivers don't
reply anything.

To be able to understand what is eventually happening, we need to
enable more log related to PNO. This Cl should help with this.

Must revert this CL once we got all the information needed.

Test: Manual
Bug: 180659832

Signed-off-by: mullerf <mullerf@google.com>
Change-Id: I49c15dddba17eefe901afc4ebe7426a61b0aa5be