Fix charging bugs
Use a headroom property to determine when we should move to the next
voltage threshold. This is because the charger already enters the CV
stage and starts dropping the current before the voltage actually
reaches the next threshold.

Also, avoid calculating charging current when we are not charging.
Similarly, only reset charging parameters when we stop charging and not
when we resume.

Bug: 181701746
Bug: 182193667
Change-Id: I539807893c5c0c3a2de6a0d6269fac3cf246883c