Update default values of sample rate and capture channel
Sample rate and capture channel have changed to 16KHz and RIGHT Channel.
The reason we need to update this here is because at the time of probing
MCU might not be up so we need to return the proper values matching the
actual MCU configuration. If this is not done and 48K, LEFT is kept as
default, everytime a recording mixer path is applied (which is set to
16KH, RIGHT), Android will revert things back to 48K, LEFT when
finished. This additional overhead is not neccessary. Also, for better
alignment between AP and MCU configuration, it's best to keep things the
same.

Test: Use MicTest app to record audio at 16KHz, Right

Bug: 181993862
Change-Id: I403f19fdc65e39dfcba2c1bfb804936a4c725bb7