[Smart Speaker] Add sample rate option for smart amp speaker
On Proto1 devices, default sample rate for UAIF1 is 16KHz. This bus
is shared between the microphone and speaker. This change allows
configures the speaker to also run at 16KHz.

On EVT+ devices, speaker lives on UAIF0 and has a default sample rate of
48KHz (since it is not sharing the bus with the microphone).

This CL allows the speaker driver to choose the proper sample rate based
on the devicetree node.

Test: Run alarm on Proto1 + EVT and confirm proper sample rate and sound
generation.

Bug: 180514440
Change-Id: I98d17a04fd718170220af8ee78005ecad6d60912