Configure sample rate based on devicetree if available
This CL sets the speaker sample rate to the value found in devicetree.

On pre-EVT devices, speaker sample rate must be configured to 16KHz
since the bus is shared with the microphone.

On EVT+ devices, speaker is on its own bus and its sample rate is set to
48KHz.

Test: On DEV2 and EVT device confirm sound played from speaker and
correct sample rate configured.

Bug: 180514440
Change-Id: I251a3b85a77e1d1a2b35f08d8c509586d04ae644