Optionallly killable freezer
Copy of the logic in https://partner-android-review.git.corp.google.com/c/kernel/private/wembley/+/1627967/

Adds a configurable option to allow killing frozen tasks in a way
similar to cgroups v2. The status of frozen tasks is changed to
TASK_INTERRUPTIBLE and their PF_FROZEN flag is reset on pending fatal
signals.

This behavior is enabled via a run-time configurable option, so that
the pre-existing behavior is preserved by default.

Bug: 143308662
Bug: 183741253
Change-Id: I203c4045739a88d48972afde8f3bad6ef61bfce6
Signed-off-by: Yeabkal Wubshit <yeabkal@google.com>