Increase kernel log buffer size
Under the v4.9 kernel, the log size was set this to 2^20 = 1 MiB but,
as part of the upgrade to v4.19, it was reduced to the 2^17 = 128 KiB
default.   Increase it back up to 2^19 = 512 KiB so the entire boot up
log can be captured and saved into the persistent logcat section of
bugreports.

Change-Id: Id3f6a03c8dda3c1a51ac08b208c4a9fe0886a7af