Only de-assert dev_wake/pm suspend device when Tx is done
Currently, dev_wake could be de-asserted before Tx is done because we
autosuspend automatically based on when Tx started. But Tx might last
longer than the autosuspend timer.

Check for the message from the HIDL to know when UART Tx is done, and
only autosuspend the devcie at that moment, which will have for effect
to only de-assert dev_wake at that moment as well.

Test: Manual
Bug: 190331425

Signed-off-by: mullerf <mullerf@google.com>
Change-Id: I918262e8a6ad9a9df8938e0e978ed228445a432e