Increase the battery and PPG NTC polling delays
With the exception of PPG_NTC on Proto devices since the location of
that one led to very fast heating. This will reduce the average rate of
NTC polling to a more reasonable one instead of twice per second.

Changing it now because the test data shows that we no longer need such
a fast polling rate, based on how fast the device actually heats up

Bug: 188443364
Change-Id: I026359e8677aaf05a3342ecc6316100c11cdd019