[9110] abox: change cpu gear sync to not sync
When ap is suspended in abox on state,
if abox enable is called and cpu_gear_sync is called,
cpu gear change func is immediately flush to flush_work,
causing a problem. When ap is suspended,
deadlock occurs because it freezes.

Resolved by changing cpu_gear_sync to cpu_gear
so it can run later in workqueue

Bug: 192299164
Signed-off-by: Seungbin Lee <seungbin.lee@samsung.com>
Change-Id: Iaef0ce0ffdd9a53620cad94f3f01725a3d95addf