Ignore packet errors during MIPI read commands
Due to the 9110 chipset hardware, it cannot do the correct ULPS exit
sequence which causes the panel to reply with an error code upon the
next MIPI read. But the correct information is still sent along with the
error code, so just ignore the error code from the MIPI read functions.

Bug: 191797968
Signed-off-by: Alice Sheng <alicesheng@google.com>
Change-Id: Id2fc9e2b9afaaa8c30e15b4edc30274f7d8b087a