Fix a race condition when enabling LPM
When LPM is enabled, the BT HAL does not send a new PROC_WRITE == '1'
before starting Tx transfer. Because we used to autosuspend right after
LPM was enabled, this was leading to the driver suspending in middle of
a Tx transfer if it was taking longer than 1 second.

Now, we don't autosuspend when enabling LPM, and rather wait for the
PROC_WRITE == '0' to let us know that the tranfer is effectively over
and that we can autosuspend there.

Test: Manual
Bug: 195019001

Signed-off-by: mullerf <mullerf@google.com>
Change-Id: I2f09d02e0bbfe91e95a4148fc0afa1f9d6aedd71