Re-enable rotating side button power management
We had previously disabled PM ops to enable wake-on-scroll. Later it was
discovered that this feature (wake-on-scroll) led to a poor user
experience and so was disabled.

Re-enabling RSB PM ops should reduce idle power.

This change also cleans up the interaction between the main thread and
the worker threads to make sure we don't have any race conditions.

Signed-off-by: Matthew Bouyack <mbouyack@google.com>
Bug: 192268197
Change-Id: Ie63a50895e5b2c24e296e1b7e59e516b1f3e6447