Nanohub flow control with exponential backoff
This CL modifies the Nanohub busy mechanism to make it suitable for use
in a general flow control scheme. This form of flow control will be
used with a new MCU traffic strategy, which generally guarantees fast,
fixed transaction timing, with backoff for overruns and other extremes.

The MCU will be configured with sufficient capacity to trigger flow
control rarely, but when it does the performance will degrade smoothly.

The exponential backoff uses sleep instead of udelay for less CPU load.
This is shown to behave just as well as udelay for the traffic patterns
of interest.

This change is important to enable SPI DMA on the AP because it can
cope with the wider range of timing that results when using DMA.

Specific use cases include:
- write overruns during MCU startup (implicated in sensor flakiness)
- abnornmal timing, eg, during display offload where heavy graphics
  initialisation can occur

See go/rh-nanohub-mcu-architecture

Bug: 200859803
Bug: 200616501
Test: prolonged stress testing with and without induced retries
Change-Id: Ic37ea88b058955f7ca5bb57ec0d18d6903da8ac7