Modify trigger setting for WiFi OOB pin from edge to level
The original vendor drop was setting the pin that way, then it got
changed when updating the code in the driver for the move from kernel
4.9 to kernel 4.19. However, if the impact isn't observed right away
and for most cases won't affect wifi, it has been observed that the OOB
pin can produce noise on the pin sometimes, which leads to it being
interpreted as an interrupt by the SoC and when it really shouldn't.
This leads to WiFi hanging and crashing. By setting the trigger to
level, it prevents this issue.

Test: Manual
Bug: 201306566
Bug: 201460624
Bug: 201565289

Signed-off-by: mullerf <mullerf@google.com>
Change-Id: I0e63232d1a117c764b9a232d77f055ae2ec09418