DMA support for Nanohub
- ensures burst mode operation for fast, continuous clock
- adjusts the timeout padding to maintain similar timing to the MCU
- relies on the new exponential back-off flow control for overall
  reliability

DMA is critically dependent on burst mode operation, and is essentially
unusable without it. Burst mode ensures a solid clock that doesn't
stutter under system load or governor control.

DMA has advantages and disadvantages:
- continuous clock in burst mode versus 64 bytes on/64 bytes off in
  non-dma mode
- less cpu load and interrupt driven without the spinlocks used in
  non-DMA mode
- slower transaction setup and hold times which limits performance,
  we will improve this further
- non-burst mode degrades markedly with mif qos so we won't be using it

The timing is set to approximate the current transaciton timing. It
will be tuned and customized for optimal timing with DMA in follow up
CLs.

See go/rh-nanohub-mcu-architecture

Test: manual, with induced retries and continuous stress tests
Bug: 185406845
Bug: 169772504
Bug: 200616501
Change-Id: I7f90f250f9c902ee473606ee2ec2383b409ec874