Adjust down charge_full_design to derated 279 mAh
Previously 300 mAh was used and while the pack itself is labeled with
a sticker saying 294 mAh, this number should be reduced to the derated
value of 279 mAh to be consistent with what we're using in modeling.

Note that within the Linux kernel battery capacity is discussed using
three different terms.

  1. "charge" represents capacity in µAh only
  2. "energy" represents capacity in µWh only
  3. "capacity" represents capacity in *percents*, from 0 to 100

Thus the fuelgauge,capacity_max property should be in percents but it
is actually in tenths of a percent.  Thus the 1000 number currently in
the device tree actually means 100%.  More problematically, this value
is improperly sent to userspace as POWER_SUPPLY_PROP_ENERGY_FULL_DESIGN,
but luckily, Android does not appear to use this power supply property.

Bug: 202396388
Signed-off-by: Tim Kryger <tkryger@google.com>
Change-Id: Ia912e5ba668b16af1703bf8adf2113abf2693378