Reduce power_supply_changed use when unnecessary
The power_supply_changed call wakes up userspace and triggers a lot of
activity. Remove cases where it's not actually useful/necessary to wake
up userspace:
- When power is connected/disconnected during WLC, as that is supposed
to happen many times normally throughout the charging cycle
- When the temp index changes, this doesn't result in any meaningful
action from userspace and may trigger many times during WLC and when the
temperature is on the border.

Bug: 202267679
Change-Id: Ifd5d071e8b15b9df7ca79884bcb13b5453a7caad