Remove monitor_wake_lock, move resume work queuing
The wake lock is no longer necessary now that we flush the work on
suspend and the wake lock is causing aborts, so remove it.

Moving the queuing of the work on resume to pm_notifier instead of
suspend complete allows us to get rid of the 200ms delay before running
it, and also allows us to, in most cases, avoid running the work on
suspend as well since that delays suspend. The 200ms delay was
originally necessary because thaw_workqueues would wait for the
bat_monitor_work in the queue to finish before continuing resume, so any
delays in bat_monitor_work would also delay resume. The PM_POST_SUSPEND
call comes after thaw_workqueues, so we don't have that issue.

Bug: 201846168
Signed-off-by: Alice Sheng <alicesheng@google.com>
Change-Id: Iff2344499151fed1dd09a59beff66cd8f9eddf70