Move nfc_active to novatek block
The code was originally copied from a previous CL that shared an
interrupt between novatek and nfc. For this particular CL, novatek is
the exclusive owner of the pin, so there is no need to put it as part of
the pinctrl.

The issue with having it part of the pinctrl is pinctrl_0 already has
the interrupt pin for haptics/smart amp, so just due to the loading
order this clobbered the haptics/smart amp interrupt causing it to be a
pull-down (on an active-low interrupt), which caused it to continuously
fire.

Bug: 203011584
Signed-off-by: Ben Fennema <fennema@google.com>
Change-Id: I67295b89f4914184885e71045c1803e94d398a74