nitrous: Remove bt_host_wake wakelock release 1s delay
The 1 second delay in releasing the wakelock held when bt_host_wake is
asserted was arbitrarily chosen initially. After investigating this and
also double checking with the BT vendor, it makes no sense keeping a
lock after the BT chip de-assert bt_host_wake. In addition, thr BT chip
already holds the bt_host_wake asserted for ~300 ms after the UART Rx
was transmitted, so that already prevents the lock to be released too
quickly.

Test: Manual
Bug: 203099428

Signed-off-by: Florian Muller <mullerf@google.com>
Change-Id: Ia840dcde7f4f1818ca9f64507ed3bb72be3e23f5