Read brightness from panel when exiting doze.
`backlight_update_status` is called when switching mode, which sends a
brightness command to the panel using previous brightness value.
However for the case of exiting doze, this value could be outdated as
the MCU might have changed it during AOD while AP was sleeping.

To make sure the brightness is in sync, retrieve the value from panel
when exit doze.

Bug: 206605823
Test: Manual
Signed-off-by: Chuong Hoang <chuongh@google.com>
Change-Id: I8fa6b5b2964390e108a727f373d64f26c63f3668