Keep S3C RTC block disabled on EVT and later
Prior to EVT / Dev3, the 32k used by haptics, mcu, and other external
ICs was sourced from the RTCCLKOUT pin associated with the RTC block
in the main part of the SoC and in order to enable the output of the
clock, the s3c-rtc driver needed to be loaded to configure registers.

However, now the 32k is being sourced from the PERI32K pin of the PMIC
such that no driver is needed and furthermore, it now appears that the
s3c-rtc driver is unexpectedly seeing interrupts and taking wakelocks
which is causing attempted entries into suspend to abort.

Since this driver is not necessary and is creating problems, keep it
from activating by marking the hardware declaration disabled within
the device tree.

Bug: 209870351
Test: No s3c2410-rtc in /proc/interrupts and 32k present at R212 on Dev3
Change-Id: I8b2ffbbde67076d150795c3cb14e324ff465515e