bcmdhd: Reduce PNO_ADAPTIVE_SCAN_LIMIT value to 30 instead of 60
Android R (11) default value is 60. So keeping this below default value
so it gives the opportunity to use the Frameworks setting for PNO scan
intervals.

Also hard-code the repeat value for PNO scan cycles to 3, since that is
what Android is using but bcmdhd drivers don't support retrieving this
directly from wificond through nl80211. And set the max backoff exp freq
to 2, as this will allow PNO scan cycles as follow:

60s -> 60s -> 60s -> 120s -> 120s -> 120s -> 240s -> [forever] -> 240s

Test: Manual
Bug: 215656628

Signed-off-by: Florian Muller <mullerf@google.com>
Change-Id: I87468b6284a4476c7e164f9304b7aed2094181d8