Revert^2 "nitrous: Remove bt_host_wake wakelock release 1s delay"
This reverts commit 52df49e5f526d2a76787992fbdd6f44a8a935e0a.

Reason for revert: After running for three weeks with this reverted on my own watch and have the test team in TW running BCST, it is clear that this change does not affect in any way the crash rate we currently observe on BT overall. It might have a positive impact on power though, so it makes sense to revert this.

Change-Id: I64831429ee0b776b1a86410a874c35f5bbd99d82