Support early-display-wakeup
Add a IRQ which will notify decon of TTW. If it fires while the device
is awake, queue a work to turn the display on. Also queue a work to turn
it back off after 1 second if the real display on signal doesn't come.

If the device is asleep, don't queue the work until the resume notifier.

Bug: 222158397
Signed-off-by: Alice Sheng <alicesheng@google.com>
Change-Id: Ieb4233d6e957d51667d104594993d2e0c2180984