Avoid accessing i2c from touch in suspend
It is suspected that a touch interrupt which occurs late in suspend
may cause us to access i2c after it has suspended. This change
disables the interrupt prior to suspend to avoid this case.

Note that even with the interrupt disabled we can still call
enable_irq_wake to allow touch events to wake the device.

Test: tap-to-wake (check 'dmesg' to confirm device was suspended)
Signed-off-by: Matthew Bouyack <mbouyack@google.com>
Bug: 217742260
Change-Id: I6d726a0c584bf82a03560d95303b8b5ad6fe6179