MCU fast boot via NXP bootloader over SPI
The Rohan MCU currently takes 30s to boot over UART. This is
unacceptable because:
- it adds 30s to the visible boot, ie, until the watch face is visible
- MCU restarts are so slow that sensor behaviour is impacted

We cannot use blhost over SPI because it requires either SPIDEV, or a
custom interface, precise timing control, and workarounds for bootloader
bugs. None of these things are viable in Rohan.

This CL adds support for a subset of the NXP bootloader commands, enough
to enable booting the MCU over SPI. Other commands can be easily added
on top of the basic interface.

- MCU boot time is improved from 30s to 2s
- bootloader sync and timing bugs are handled
- bootloader data aborts are detected and retried
- downloads run at elevated priority for speed and stability
- signed firmware is supported

See go/rh-mcu-boot-improvements

Bug: 206044802
Change-Id: I79b9d9d7ab670e1e758824dfaf212b46e41120c2