of: property: Add "depends-on" support
This is a hack to use on this specific kernel version and this specific
product. This will not be supported on future kernel versions. For
future kernel versions, the dependencies should be captured properly in
DT using upstream supported DT properties/bindings.

Add support for a depends-on property that teaches of_devlink that the
node containing the property depends on the node pointed to by the
property.

Bug: 196103381
Signed-off-by: Saravana Kannan <saravanak@google.com>
Change-Id: I0cdddf8d58fcb0bf9f43be77060d558c6b660fdf