nitrous: Add a wakelock to manage the UART Tx transactions
We are seeing issues where it seems that the userspace BT wakelocks
from the stack and from the HAL are failing to get acquired when Tx
UART data are sent. It leads a different type of issues but so far,
the issues identified are either crashes or UART BT irq stuck in a
loop.
As a workaround to this issue, we acquire a lock in the kernel at the
start of a UART Tx request and release it at then end with a 100 ms
timer to match what the BT stack and HAL usually do. This is with the
hope that kernel locks acquired in the kernel will not fail to abort a
suspend even if already advanced in that stage, which seems to be where
the usersapce locks are failing.

Test: Manual
Bug: 226972758
Bug: 195363858

Signed-off-by: Florian Muller <mullerf@google.com>
Change-Id: Ib2d2434ef0557d7f707b973cb88bf30ca4b9cc24