Reduce empty packet log spam
If the MCU becomes stuck, or is slow to respond, with the data interrupt
line low, the AP will attempt many retries trying to read the data,
which creates a lot of log spam since we are running with debug logs
enabled for nanohub.

Here we reduce the log of empty ack packets to a single tally of the
total number of empty acks. Even with this there can be 10 outer
retries, so this message can itself be repeated 10 times if the MCU
is stuck.

Bug: 228919054
Test: Manually induced retries and observed logs
Change-Id: I5d6e295f612b6f367726ae41844c1d8d3f48081f