Use a double pulse when resetting the MCU to avoid boot failure
This is a workaroud for:
b/227924872 R11 MCU boot failure (due to SPI1, JTAG and RESET interaction)

After multiple resets, there is a chance the MCU can enter a TAP test
mode that it cannot escape from due to interaction b/w the SPI1 bus and
the JTAG control pins. The ultimate cause is a lack of pull-down on the
TRST_N pin.

A double reset pulse works around this problem as follows:
1. The first reset ensures the TRST_N pin is deterministically pulled
low by the bootrom.
2. The second reset occurs while TRST_N is still low, which will in
turn reset the TAP state machine.
This ensures that the TAP state machine cannot cumulatively move to a
dangerous state across multiple MCU boots.

Bug: 227924872
Test: Extensive soak testing with 1000s of reboot cycles with no failures
Change-Id: I8a79e4e90e7df02657d9a40f50bbf5b3d4fc2799