Address race condition around early ttw queueing
The TTW can queue in between when the cancel_delayed_work_sync executing
and the display actually changing to the ON state during the UNBLANK
ioctl.

In between the UNBLANK and POWERDOWN FB messages, we don't want any of
the early TTW related logic to be able to queue or execute, so block it
during that time.

Bug: 230648842
Change-Id: I271ae47bf2baeb07ba4be76acb80844ca8d0cffb
Signed-off-by: Alice Sheng <alicesheng@google.com>