Ensure the SPI bus idle state is correct after suspend/resume
The SPI driver does not always preserve or restore the bus idle state
after suspend/resume. Eg, when using CPHA, the clock should idle high,
but it often goes low in suspend and does not return high. I looked for
DT-based ways to fix this, but the driver is hardcoded and the only
soluion is the one here, or to cede control of chip-select back to the
SPI driver.

Bug: 233382308
Test: extensive manual verification and testing
Change-Id: I7a871deca9c65694223d535dfec538856deb3a8f