Don't reset the MCU on nanohub errors
We have observed in b/231130162 that nanohub can reset the MCU shortly
after a firmware download. Whilst we rely on the MCU watchdog to catch
and handle any and all cases like this, this is not supposed to happen
and it can interfere with the MCU watchdog, restarts and crash
reporting. It can also lead to errors in telemetry. Although resets
were being logged in the past, they weren't actually being performed,
so here we also change the log to be more accurate.

Bug: 231130162
Test: manually induced errors and verified behaviour
Change-Id: I232dc218f3131cac3140a5ee63b8a84c86fd2a42