abox: Fix IPC related issues
1. Update Calliope state to CALLIOPE_ENABLING as soon as the work to
download the abox FW has started. Otherwise we might run into a race
condition where the state is still CALLIOPE_DISABLED, which wouldn't be
expected here and would lead to an error.
2. Check for abox boot done (check for Calliope state CALLIOPE_ENABLING
vs CALLIOPE_ENABLED) when sending IPC and behave accordingly depending
of the state in order to prevent "Transaction timeout" issue with abox.

Test: Manual
Bug: 238920301
Bug: 238470383
Signed-off-by: Florian Muller <mullerf@google.com>
Change-Id: I7073991656559a45e6c56ac6c2d134cc7fd3217b