Fix unintended sign extension in RSB driver
The magnitude of RSB rotation is a 12-bit value which is read from the
hardware as 8 low bit and 4 high bits. Because the 8 low bits were
being stored in an int8_t, casting that value to int16_t was producing
unintended sign extension. This caused very large values to become
negative.

Also, there is no reason to cast the rotation magnitude to 's8' before
calling input_report_rel. That API accepts a 32-bit integer.
This was also causing issues with overflow.

Signed-off-by: Matthew Bouyack <mbouyack@google.com>
Bug: 234361169
Change-Id: I3ea53ffd392a11ecc6c8bb6ac8cca49d54fc0df7