abox: Grab a wakelock every time we are in MODE_IN_CALL
This is to workaround the fact that there is a noticeable decrease in
quality of the uplink every time the SoC goes to suspend during an HFP
call.
So here, we grab a lock while R11 is the active device for the call,
and release the lock in all other cases. Note that this lock affects
both HFP and modem call cases. But the modem call case already uses a
lock managed by the Dialer app, so it effectively does not change the
behavior there.

Test: Manual
Bug: 231630258
Bug: 224400455
Bug: 239966045
Signed-off-by: Florian Muller <mullerf@google.com>
Change-Id: I2bf952616121464fa6260ea78dd559db9bf2974d