Set client->irq prior to registering touch input device
Previously we were registering the touch panel as an input device
before initialization was actually complete. In particular, we
had not yet initialized the touch interrupt handler.

This was causing issues in the case where nvt_ts_input_open is
called immediately after the panel is registered but while
client->irq is still zero.

Signed-off-by: Matthew Bouyack <mbouyack@google.com>
Bug: 249214710
Change-Id: I00073646be44527cd974efa80538c0b222a609d4