i2c: exynos5: correct 0-length read behavior
Adapted from pa/2190845
  - Small code style changes

Modify the driver to check the buffer length
*prior* to writing into the buffer, on i2c
reads.

this corrects a verified issue where, if a
zero-length read is issued and an IRQ fires,
the driver will presently do an out-of-bounds
write.

Also update the fifo underrun trigger levels
to have a lower bound of the bus width,
as required by the datasheet.  This might
prevent the spurious IRQ on 0-length reads.

Bug: 195480799
Bug: 253010037
Signed-off-by: Necip Fazil Yildiran <necip@google.com>
Change-Id: I20a178c81223444ef17d97a7570bf3c46d800546