soc/google/cpif: Fix use-after-free caused by cpif module
During the cpif probe function, various notifier blocks are registered
and, if the cpif probe function fails at any stage, an error is
reported, and none of the notifiers are unregistered.

This is problematic as the notifier blocks are part of structures
that are deallocated after the cpif probe function fails. This allows
any interaction with a notifier chain that holds a reference to these
structures (e.g. pm_notifier chain) to refer to memory that has
been freed when accessing the notifier chain.

Add error handling for cases where the cpif probe function, fails so that
it properly cleans up any resources that it allocates.

This patch has been ported from the original patch pa/2332137. Please
check the bug for KASAN report.

Bug: 261647122
Change-Id: Ie53d5efd607634a4a241a847898f054832b41ade
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>
Signed-off-by: Mahesh Kallelil <kallelil@google.com>