abox: Try to recover from an out of sync usage_count
When the usage_count gets out of sync, the system stops suspending,
which leads to intense battery drain.
In an attempt to get back to normal, try to recover from this situation
with the last resort being to reset the usage_count and allow suspend.

Test: Manual - tested multiple audio cases while taking the device out
of suspend and also letting it suspend, then verified bugreport to make
sure the driver behaved as expected.
Bug: 242936355
Bug: 243041824
Change-Id: Idcb9d18442407e6fe4c77e5c17a9acae8e6a4cf5
Signed-off-by: Florian Muller <mullerf@google.com>